🎊📫 Checkpoint ./komutlar/iletişim.js:14029226/6 ./komutlar/afk.js:14029226/1783 ./Strom.js:14029226/2365

+++ Strom.js
@@ -602,41 +602,49 @@ global.onUpdateInvite = (guildMember, guild, total) => {
 
 
 
-//     [-----------------> Afk <------------------]  \\
+//AFK Baş
+
+const ms = require("parse-ms");
+const { DiscordAPIError } = require("discord.js");
 
 client.on("message", async message => {
-  
-  let prefix = ayarlar.prefix;
-  let kullanıcı = message.mentions.users.first() || message.author;
-  let afkdkullanıcı = await db.fetch(`afk_${message.author.id}`);
-  let afkkullanıcı = await db.fetch(`afk_${kullanıcı.id}`);
-  let sebep = afkkullanıcı;
   if (message.author.bot) return;
-  if (message.content.includes(`${prefix}afk`)) return;
-  if (message.content.includes(`<@${kullanıcı.id}>`)) {
-    if (afkdkullanıcı) {
-      message.channel.send(
-        new Strom.MessageEmbed().setDescription(`
-AFK modundan ayrıldın <@${kullanıcı.id}>.`)
-      );
-      db.delete(`afk_${message.author.id}`);
-    }
-    if (afkkullanıcı)
-      return message.channel.send(
-        `${message.author}\`${kullanıcı.tag}\` şu anda AFK. \n Sebep : \`${sebep}\``
-      );
+  if (!message.guild) return;
+  if (message.content.includes(`afk`)) return;
+
+  if (await db.fetch(`afk_${message.author.id}`)) {
+    db.delete(`afk_${message.author.id}`);
+    db.delete(`afk_süre_${message.author.id}`);
+
+    const embed = new Discord.MessageEmbed()
+
+      .setColor("#00ff00")
+      .setAuthor(message.author.username, message.author.avatarURL)
+      .setDescription(`${message.author.username} Artık \`AFK\` Değilsin.`);
+
+    message.channel.send(embed);
   }
-  if (!message.content.includes(`<@${kullanıcı.id}>`)) {
-    if (afkdkullanıcı) {
-      message.channel.send(
-        new Strom.MessageEmbed().setDescription(`
-AFK modundan ayrıldın <@${kullanıcı.id}>.`)
+
+  var USER = message.mentions.users.first();
+  if (!USER) return;
+  var REASON = await db.fetch(`afk_${USER.id}`);
+
+  if (REASON) {
+    let süre = await db.fetch(`afk_süre_${USER.id}`);
+    let timeObj = ms(Date.now() - süre);
+
+    const afk = new Discord.MessageEmbed()
+
+      .setColor("#00ff00")
+      .setDescription(
+        `**Bu Kullanıcı AFK**\n\n**Afk Olan Kullanıcı :** \`${USER.tag}\`\n**Afk Süresi :** \`${timeObj.hours}saat\` \`${timeObj.minutes}dakika\` \`${timeObj.seconds}saniye\`\n**Sebep :** \`${REASON}\``
       );
-      db.delete(`afk_${message.author.id}`);
-    }
+
+    message.channel.send(afk);
   }
 });
 
+//AFK Son
 
 
 ///reklam engel
